Prevent errors from occurring when the message data is malformed.
On some builds the dice roll will fail due to the messageData object not having the update method. This will cause the dice roll to fail without any information being passed back to the end user. I found on the latest V9 update, this issue was becoming quite prevalent as my world and a friendly world encountered it. This change makes it so even if it fails, the user will get a console log to see what went wrong.