adding degenesis system
besides the basic system config...

added a degenesis specific html file. The existing one had a bunch of fields that aren't relevant to the degenesis system. The skills were also overflowing off the sheet because of many skills with longer names

had to create LMRTFY.skills by reading from an actor. I had to do this because this type of system info wasn't readily available without importing a file and even then it wouldn't be in the proper format to work with LMRTFY