Refactor: Split RollProviders
This new setup splits out roll providers. How this works is that each system operates its own functions. Rather than having systems ingrained into the code itself, each of the systems determine if they need to override specific parts of the rolling system.

This will reduce the need to check in changes to the main rolling system and allow providers to be more easily created. This also means that every system is isolated in its changes to the main rolling system.